3rd commit

+++ app/views/category/index.php
@@ -10,7 +10,11 @@
 
 <body>
     <h1>Categories</h1>
-    <a href="<?php echo URLROOT; ?>/categories/add"><button>Add New Category</button></a>&nbsp;
+    
+    <?php if ($_SESSION['role'] === 'admin'): ?>
+        <a href="<?php echo URLROOT; ?>/categories/add"><button>Add New Category</button></a>&nbsp;
+    <?php endif; ?>
+    
     <a href="<?php echo URLROOT; ?>/products"><button>Go to products</button></a>&nbsp;
     <a href="<?php echo URLROOT; ?>/choose/options"><button>Home</button></a><br>&nbsp;
     <br>
@@ -27,8 +31,10 @@
                     <td style="text-align: center;"><?php echo htmlspecialchars($category->id); ?></td>
                     <td style="text-align: center;"><?php echo htmlspecialchars($category->categoryName); ?></td>
                     <td style="text-align: center;">&nbsp;
-                        <a href="<?php echo URLROOT; ?>/categories/edit/<?php echo htmlspecialchars($category->id); ?>" style="text-decoration:none;"><button>Edit</button></a>&nbsp;
-                        <a href="<?php echo URLROOT; ?>/categories/delete/<?php echo htmlspecialchars($category->id); ?>" style="text-decoration:none;" onclick="return confirm('Are you sure you want to delete this category?');"><button>Delete</button></a>&nbsp;
+                        <?php if ($_SESSION['role'] === 'admin'): ?>
+                            <a href="<?php echo URLROOT; ?>/categories/edit/<?php echo htmlspecialchars($category->id); ?>" style="text-decoration:none;"><button>Edit</button></a>&nbsp;
+                            <a href="<?php echo URLROOT; ?>/categories/delete/<?php echo htmlspecialchars($category->id); ?>" style="text-decoration:none;" onclick="return confirm('Are you sure you want to delete this category?');"><button>Delete</button></a>&nbsp;
+                        <?php endif; ?>
                         <a href="<?php echo URLROOT; ?>/categories/show/<?php echo htmlspecialchars($category->id); ?>" style="text-decoration:none;"><button>View</button></a>&nbsp;
                     </td>
                 </tr>
